refac(rend/saga): Rewrite sagas in TS
